fix: refac library
