+++ client/src/pages/trainer/trainer-library/trainer-library.module.scss
@@ -5,9 +5,6 @@
 }
 
 .header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
     margin-bottom: 2rem;
 }
 
@@ -15,7 +12,7 @@
     font-size: 2rem;
     font-weight: 700;
     color: #1a1a1a;
-    margin: 0;
+    margin-bottom: 2rem;
 }
 
 .backLink {
